x86_64 compatiablity for P.E.Op.S OpenGL Driver (untested)
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@23601 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/peopsxgl/draw.c
@@ -1018,14 +1018,14 @@ void offsetST(void)
 
 ///////////////////////////////////////////////////////// 
 
-void offsetScreenUpload(long Position)
+void offsetScreenUpload(int Position)
 {
  if(bDisplayNotSet)
   SetOGLDisplaySettings(1);
 
  if(Position==-1)
   {
-   long lmdx,lmdy;
+   int lmdx,lmdy;
 
    lmdx=xrUploadArea.x0;
    lmdy=xrUploadArea.y0;
