+++ plugins/peopsxgl/draw.h
@@ -36,7 +36,7 @@ BOOL offset4(void);
 BOOL offsetline(void);
 void offsetST(void);
 void offsetBlk(void);
-void offsetScreenUpload(long Position);
+void offsetScreenUpload(int Position);
 void assignTexture3(void);
 void assignTexture4(void);
 void assignTextureSprite(void);
