+++ plugins/peopsxgl/texture.h
@@ -19,42 +19,41 @@
 #ifndef _GPU_TEXTURE_H_
 #define _GPU_TEXTURE_H_
 
-#define TEXTUREPAGESIZE 256*256
+#define TEXTUREPAGESIZE 256 * 256
 
 void           InitializeTextureStore();
 void           CleanupTextureStore();
-GLuint         LoadTextureWnd(long pageid,long TextureMode,unsigned long GivenClutId);
+GLuint         LoadTextureWnd(int pageid, int TextureMode, uint32_t GivenClutId);
 GLuint         LoadTextureMovie(void);
-void           InvalidateTextureArea(long imageX0,long imageY0,long imageX1,long imageY1);
+void           InvalidateTextureArea(int imageX0, int imageY0, int imageX1, int imageY1);
 void           InvalidateTextureAreaEx(void);
 void           LoadTexturePage(int pageid, int mode, short cx, short cy);
 void           ResetTextureArea(BOOL bDelTex);
-GLuint         SelectSubTextureS(long TextureMode, unsigned long GivenClutId);
+GLuint         SelectSubTextureS(int TextureMode, uint32_t GivenClutId);
 void           CheckTextureMemory(void);
-
 
 void           LoadSubTexturePage(int pageid, int mode, short cx, short cy);
 void           LoadSubTexturePageSort(int pageid, int mode, short cx, short cy);
 void           LoadPackedSubTexturePage(int pageid, int mode, short cx, short cy);
 void           LoadPackedSubTexturePageSort(int pageid, int mode, short cx, short cy);
-unsigned long  XP8RGBA(unsigned long BGR);
-unsigned long  XP8RGBAEx(unsigned long BGR);
-unsigned long  XP8RGBA_0(unsigned long BGR);
-unsigned long  XP8RGBAEx_0(unsigned long BGR);
-unsigned long  XP8BGRA_0(unsigned long BGR);
-unsigned long  XP8BGRAEx_0(unsigned long BGR);
-unsigned long  XP8RGBA_1(unsigned long BGR);
-unsigned long  XP8RGBAEx_1(unsigned long BGR);
-unsigned long  XP8BGRA_1(unsigned long BGR);
-unsigned long  XP8BGRAEx_1(unsigned long BGR);
-unsigned long  P8RGBA(unsigned long BGR);
-unsigned long  P8BGRA(unsigned long BGR);
-unsigned long  CP8RGBA_0(unsigned long BGR);
-unsigned long  CP8RGBAEx_0(unsigned long BGR);
-unsigned long  CP8BGRA_0(unsigned long BGR);
-unsigned long  CP8BGRAEx_0(unsigned long BGR);
-unsigned long  CP8RGBA(unsigned long BGR);
-unsigned long  CP8RGBAEx(unsigned long BGR);
+uint32_t       XP8RGBA(uint32_t BGR);
+uint32_t       XP8RGBAEx(uint32_t BGR);
+uint32_t       XP8RGBA_0(uint32_t BGR);
+uint32_t       XP8RGBAEx_0(uint32_t BGR);
+uint32_t       XP8BGRA_0(uint32_t BGR);
+uint32_t       XP8BGRAEx_0(uint32_t BGR);
+uint32_t       XP8RGBA_1(uint32_t BGR);
+uint32_t       XP8RGBAEx_1(uint32_t BGR);
+uint32_t       XP8BGRA_1(uint32_t BGR);
+uint32_t       XP8BGRAEx_1(uint32_t BGR);
+uint32_t       P8RGBA(uint32_t BGR);
+uint32_t       P8BGRA(uint32_t BGR);
+uint32_t       CP8RGBA_0(uint32_t BGR);
+uint32_t       CP8RGBAEx_0(uint32_t BGR);
+uint32_t       CP8BGRA_0(uint32_t BGR);
+uint32_t       CP8BGRAEx_0(uint32_t BGR);
+uint32_t       CP8RGBA(uint32_t BGR);
+uint32_t       CP8RGBAEx(uint32_t BGR);
 unsigned short XP5RGBA (unsigned short BGR);
 unsigned short XP5RGBA_0 (unsigned short BGR);
 unsigned short XP5RGBA_1 (unsigned short BGR);
